+++ code/backend/Cleanuparr.Api/Features/General/Contracts/Requests/UpdateGeneralConfigRequest.cs
@@ -2,7 +2,6 @@ using Cleanuparr.Domain.Enums;
 using Cleanuparr.Infrastructure.Http.DynamicHttpClientSystem;
 using Cleanuparr.Infrastructure.Logging;
 using Cleanuparr.Persistence.Models.Configuration.General;
-using Serilog.Events;
 
 namespace Cleanuparr.Api.Features.General.Contracts.Requests;
 
@@ -82,66 +81,5 @@ public sealed record UpdateGeneralConfigRequest
 
         logger.LogCritical("Reconfiguring logger due to configuration changes");
         LoggingConfigManager.ReconfigureLogging(config);
-    }
-}
-
-public sealed record UpdateLoggingConfigRequest
-{
-    public LogEventLevel Level { get; init; } = LogEventLevel.Information;
-
-    public ushort RollingSizeMB { get; init; } = 10;
-
-    public ushort RetainedFileCount { get; init; } = 5;
-
-    public ushort TimeLimitHours { get; init; } = 24;
-
-    public bool ArchiveEnabled { get; init; } = true;
-
-    public ushort ArchiveRetainedCount { get; init; } = 60;
-
-    public ushort ArchiveTimeLimitHours { get; init; } = 24 * 30;
-
-    public bool ApplyTo(LoggingConfig existingConfig)
-    {
-        bool levelChanged = existingConfig.Level != Level;
-        bool otherPropertiesChanged =
-            existingConfig.RollingSizeMB != RollingSizeMB ||
-            existingConfig.RetainedFileCount != RetainedFileCount ||
-            existingConfig.TimeLimitHours != TimeLimitHours ||
-            existingConfig.ArchiveEnabled != ArchiveEnabled ||
-            existingConfig.ArchiveRetainedCount != ArchiveRetainedCount ||
-            existingConfig.ArchiveTimeLimitHours != ArchiveTimeLimitHours;
-
-        existingConfig.Level = Level;
-        existingConfig.RollingSizeMB = RollingSizeMB;
-        existingConfig.RetainedFileCount = RetainedFileCount;
-        existingConfig.TimeLimitHours = TimeLimitHours;
-        existingConfig.ArchiveEnabled = ArchiveEnabled;
-        existingConfig.ArchiveRetainedCount = ArchiveRetainedCount;
-        existingConfig.ArchiveTimeLimitHours = ArchiveTimeLimitHours;
-
-        existingConfig.Validate();
-
-        LevelOnlyChange = levelChanged && !otherPropertiesChanged;
-
-        return levelChanged || otherPropertiesChanged;
-    }
-
-    public bool LevelOnlyChange { get; private set; }
-}
-
-public sealed record UpdateAuthConfigRequest
-{
-    public bool DisableAuthForLocalAddresses { get; init; }
-
-    public bool TrustForwardedHeaders { get; init; }
-
-    public List<string> TrustedNetworks { get; init; } = [];
-
-    public void ApplyTo(AuthConfig existingConfig)
-    {
-        existingConfig.DisableAuthForLocalAddresses = DisableAuthForLocalAddresses;
-        existingConfig.TrustForwardedHeaders = TrustForwardedHeaders;
-        existingConfig.TrustedNetworks = TrustedNetworks;
     }
 }
